webapp/requirements.txt
bcrypt
Flask
Flask-Bcrypt
Flask-MySQLdb
Flask-SQLAlchemy
Flask-UUID
mysqlclient
PyMySQL
python-dotenv
SQLAlchemy
requests
flask-sqlalchemy
pytest
Flask-Testing
python-json-logger
pytz
google-api-python-client
google-cloud
google-cloud-pubsub